infra/collectors/requirements.txt
google-cloud-bigquery==3.27.0
google-auth==2.37.0
